+++ customer/lib/screens/home_screens/bookmark_screen.dart
@@ -13,17 +13,13 @@ class _BookmarkScreenState extends State<BookmarkScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.background,
+      appBar: AppBar(
+        title: const Center(child: Text('즐겨찾기')),
+      ),
       body: Container(
         margin: const EdgeInsets.all(20),
         child: const Column(
           children: [
-            Center(
-              child: Text(
-                '즐겨찾기 탭',
-                style: TextStyle(fontSize: 40),
-              ),
-            ),
             FavoritesWidget(
               salesStatus: '판매중',
               favoriteRestaurant: '식당314',
