+++ customer/lib/screens/home_screens/map_screen.dart
@@ -116,7 +116,6 @@ class NaverMapApp extends StatelessWidget {
             consumeSymbolTapEvents: false, // 심볼 탭 이벤트 소비 여부 설정
           ),
           onMapReady: (controller) async {
-            // 지도 준비 완료 시 호출되는 콜백 함수
             mapControllerCompleter.complete(controller);
           },
         ),
